Factor json parser to a separate object to further separate concern
